+++ CleverPush/Source/CPAppBannerModule.h
@@ -31,7 +31,7 @@
 + (void)getBanners:(NSString*)channelId bannerId:(NSString*)bannerId notificationId:(NSString*)notificationId completion:(void(^)(NSMutableArray<CPAppBanner*>*))callback;
 + (void)presentAppBanner:(UIViewController*)controller banner:(CPAppBanner*)banner;
 + (void)initSession:(NSString*)channelId afterInit:(BOOL)afterInit;
-+ (void)triggerEvent:(NSString *)key value:(NSString *)value;
++ (void)triggerEvent:(NSString *)eventId properties:(NSDictionary *)properties;
 + (void)disableBanners;
 + (void)enableBanners;
 + (void)setTrackingEnabled:(BOOL)enabled;