+++ CleverPush/Source/CPAppBannerModuleInstance.h
@@ -30,7 +30,7 @@
 - (long)getSessions;
 - (void)saveSessions;
 - (void)setBannerOpenedCallback:(CPAppBannerActionBlock)callback;
-- (void)triggerEvent:(NSString *)key value:(NSString *)value;
+- (void)triggerEvent:(NSString *)eventId properties:(NSDictionary *)properties;
 - (void)showBanner:(NSString*)channelId bannerId:(NSString*)bannerId;
 - (void)showBanner:(NSString*)channelId bannerId:(NSString*)bannerId notificationId:(NSString*)notificationId;
 - (void)startup;