+++ CleverPush/Source/CPAppBannerTriggerCondition.h
@@ -1,12 +1,13 @@
 #import <Foundation/Foundation.h>
 #import "CPAppBannerTriggerConditionType.h"
+#import "CPAppBannerTriggerConditionEventProperty.h"
 
 @interface CPAppBannerTriggerCondition : NSObject
 
 #pragma mark - Class Variables
 @property (nonatomic) CPAppBannerTriggerConditionType type;
-@property (nonatomic, strong) NSString *key;
-@property (nonatomic, strong) NSString *value;
+@property (nonatomic, strong) NSString *event;
+@property (nonatomic, strong) NSMutableArray<CPAppBannerTriggerConditionEventProperty*> *eventProperties;
 @property (nonatomic, strong) NSString *relation;
 @property (nonatomic) int sessions;
 @property (nonatomic) int seconds;